Separate global aggs from viewport aggs
