+++ src/core/viz/expressions/viewportAggregation.js
@@ -166,9 +166,6 @@ function genViewportAgg(metadataPropertyName, zeroFn, accumFn, resolveFn) {
             this.property._compile(metadata);
             this.type = 'number';
             super.inlineMaker = inline => inline.value;
-            if (global) {
-                this.value.expr = metadata.columns.find(c => c.name === this.property.name)[metadataPropertyName];
-            }
         }
         _getMinimumNeededSchema() {
             return this.property._getMinimumNeededSchema();
@@ -184,11 +181,7 @@ function genViewportAgg(metadataPropertyName, zeroFn, accumFn, resolveFn) {
             super._preDraw(...args);
         }
         _getDrawMetadataRequirements() {
-            if (!global) {
-                return { columns: [this._getColumnName()] };
-            } else {
-                return { columns: [] };
-            }
+            return { columns: [this._getColumnName()] };
         }
         _getColumnName() {
             if (this.property.aggName) {
@@ -200,160 +193,7 @@ function genViewportAgg(metadataPropertyName, zeroFn, accumFn, resolveFn) {
     };
 }
 
-
-/**
- * Return the average value of all the features
- *
- * @param {carto.expressions.Base} property - Column of the table
- * @return {carto.expressions.Base} Result of the aggregation
- *
- * @example
- * const s = carto.expressions;
- * const $amount = s.prop('amount');
- * const viz = new carto.Viz({
- *   @g_avg: s.globalAvg($amount)
- * });
- *
- * @memberof carto.expressions
- * @name globalAvg
- * @function
- * @api
- */
-export const GlobalAvg = generateGlobalAggregattion('avg');
-
-/**
- * Return the maximum value of all the features
- *
- * @param {carto.expressions.Base} property - Column of the table
- * @return {carto.expressions.Base} Result of the aggregation
- *
- * @example
- * const s = carto.expressions;
- * const $amount = s.prop('amount');
- * const viz = new carto.Viz({
- *   @g_max: s.globalMax($amount)
- * });
- *
- * @memberof carto.expressions
- * @name globalMax
- * @function
- * @api
- */
-export const GlobalMax = generateGlobalAggregattion('max');
-
-/**
- * Return the minimum value of all the features
- *
- * @param {carto.expressions.Base} property - Column of the table
- * @return {carto.expressions.Base} Result of the aggregation
- *
- * @example
- * const s = carto.expressions;
- * const $amount = s.prop('amount');
- * const viz = new carto.Viz({
- *   @g_min: s.globalMin($amount)
- * });
- *
- * @memberof carto.expressions
- * @name globalMin
- * @function
- * @api
- */
-export const GlobalMin = generateGlobalAggregattion('min');
-
-/**
- * Return the sum of the values of all the features
- *
- * @param {carto.expressions.Base} property - Column of the table
- * @return {carto.expressions.Base} Result of the aggregation
- *
- * @example
- * const s = carto.expressions;
- * const $amount = s.prop('amount');
- * const viz = new carto.Viz({
- *   @g_sum: s.globalSum($amount)
- * });
- *
- * @memberof carto.expressions
- * @name globalSum
- * @function
- * @api
- */
-export const GlobalSum = generateGlobalAggregattion('sum');
-
-/**
- * Return the count of all the features
- *
- * @param {carto.expressions.Base} property - Column of the table
- * @return {carto.expressions.Base} Result of the aggregation
- *
- * @example
- * const s = carto.expressions;
- * const $amount = s.prop('amount');
- * const viz = new carto.Viz({
- *   @g_count: s.globalCount($amount)
- * });
- *
- * @memberof carto.expressions
- * @name globalCount
- * @function
- * @api
- */
-export const GlobalCount = generateGlobalAggregattion('count');
-
-/**
- * Return the percentile of all the features
- *
- * @param {carto.expressions.Base} property - Column of the table
- * @return {carto.expressions.Base} Result of the aggregation
- *
- * @example
- * const s = carto.expressions;
- * const $amount = s.prop('amount');
- * const viz = new carto.Viz({
- *   @g_percentile: s.globalPercentile($amount)
- * });
- *
- * @memberof carto.expressions
- * @name globalPercentile
- * @function
- * @api
- */
-export const GlobalPercentile = generatePercentile(true);
-function generateGlobalAggregattion(metadataPropertyName) {
-    return class Aggregattion extends BaseExpression {
-        /**
-         * @param {*} property
-         */
-        constructor(property) {
-            super({ value: number(0) });
-            this.property = implicitCast(property);
-        }
-        eval() {
-            return this.value.expr;
-        }
-        _compile(metadata) {
-            super._compile(metadata);
-            // TODO improve type check
-            this.property._compile(metadata);
-            this.type = 'number';
-            super.inlineMaker = inline => inline.value;
-            this.value.expr = metadata.columns.find(c => c.name === this.property.name)[metadataPropertyName];
-        }
-        _getMinimumNeededSchema() {
-            return this.property._getMinimumNeededSchema();
-        }
-        _getColumnName() {
-            if (this.property.aggName) {
-                // Property has aggregation
-                return schema.column.aggColumn(this.property.name, this.property.aggName);
-            }
-            return this.property.name;
-        }
-    };
-}
-
-function generatePercentile(global) {
+function generatePercentile() {
     return class Percentile extends BaseExpression {
         /**
          * @param {*} property
@@ -375,38 +215,26 @@ function generatePercentile(global) {
             this.property._compile(metadata);
             this.type = 'number';
             super.inlineMaker = inline => inline.value;
-            if (global) {
-                const copy = metadata.sample.map(s => s[this.property.name]);
-                copy.sort((x, y) => x - y);
-                const p = this.percentile / 100;
-                this.value.expr = copy[Math.floor(p * copy.length)];
-            }
         }
         _getMinimumNeededSchema() {
             return this.property._getMinimumNeededSchema();
         }
         _getDrawMetadataRequirements() {
-            if (!global) {
-                return { columns: [this._getColumnName()] };
-            } else {
-                return { columns: [] };
-            }
+            return { columns: [this._getColumnName()] };
         }
         _preDraw(program, drawMetadata, gl) {
             // TODO use _updateDrawMetadata
             const name = this._getColumnName();
-            if (!global) {
-                const column = drawMetadata.columns.find(c => c.name === name);
-                const total = column.accumHistogram[column.histogramBuckets - 1];
-                // TODO OPT: this could be faster with binary search
-                for (var i = 0; i < column.histogramBuckets; i++) {
-                    if (column.accumHistogram[i] >= this.percentile / 100 * total) {
-                        break;
-                    }
+            const column = drawMetadata.columns.find(c => c.name === name);
+            const total = column.accumHistogram[column.histogramBuckets - 1];
+            // TODO OPT: this could be faster with binary search
+            for (var i = 0; i < column.histogramBuckets; i++) {
+                if (column.accumHistogram[i] >= this.percentile / 100 * total) {
+                    break;
                 }
-                const br = i / column.histogramBuckets * (column.max - column.min) + column.min;
-                this.value.expr = br;
             }
+            const br = i / column.histogramBuckets * (column.max - column.min) + column.min;
+            this.value.expr = br;
             this.value._preDraw(program, drawMetadata, gl);
         }
         _getColumnName() {
